extract jit functions to their own module, and make run_jit public separate from the driver
